+++ WhistleWindLobotomyMod/Cards/apocalypseBirdMinions.cs
@@ -14,7 +14,6 @@ namespace WhistleWindLobotomyMod
             const string forestKeeper = "forestKeeper";
             const string runawayBird = "runawayBird";
 
-            // spam card
             NewCard(eyeballChick, "Eyeball Chick",
                 attack: 2, health: 2)
                 .SetPortraits(eyeballChick)
@@ -23,20 +22,18 @@ namespace WhistleWindLobotomyMod
                 .AddAppearances(ForcedEmission.appearance)
                 .Build();
 
-            // only 3
             NewCard(forestKeeper, "Keeper of the Black Forest",
-                attack: 1, health: 5)
+                attack: 1, health: 4)
                 .SetPortraits(forestKeeper)
-                .AddAbilities(Ability.Sniper)
+                .AddAbilities(Ability.GainAttackOnKill)
                 .AddTribes(Tribe.Bird)
                 .AddAppearances(ForcedEmission.appearance)
                 .Build();
 
-            // regular
             NewCard(runawayBird, "Runaway Bird",
                 attack: 1, health: 3)
                 .SetPortraits(runawayBird)
-                .AddAbilities(Cycler.ability, HighStrung.ability)
+                .AddAbilities(Ability.Sniper, HighStrung.ability)
                 .AddTribes(Tribe.Bird)
                 .AddAppearances(ForcedEmission.appearance)
                 .Build();